# specs/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Krait API Documentation",
    "title": "Krait API",
    "contact": {
      "name": "Styvane Soukossi",
      "email": "mstyvane@gmail.com"
    },
    "version": "v1"
  },
  "basePath": "/v1",
  "paths": {
    "/auth/login": {
      "post": {
        "description": "Login user",
        "operationId": "Login",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "PhoneNumber",
            "description": "Phone number",
            "name": "phoneNumber",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "CountryCode",
            "description": "Country code",
            "name": "countryCode",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "Password",
            "description": "Password",
            "name": "password",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/okloginResponseWrapper"
          },
          "400": {
            "$ref": "#/responses/badLoginResponseWrapper"
          },
          "422": {
            "$ref": "#/responses/unprocessableLoginResponseWrapper"
          }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "description": "Inititate registration for new user",
        "operationId": "SignUp",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "PhoneNumber",
            "description": "Phone number",
            "name": "phoneNumber",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "CountryCode",
            "description": "Country code",
            "name": "countryCode",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/okResponseWrapper"
          },
          "400": {
            "$ref": "#/responses/badRequestResponseWrapper"
          },
          "422": {
            "$ref": "#/responses/unprocessableEntityResponseWrapper"
          }
        }
      }
    }
  },
  "responses": {
    "badLoginResponseWrapper": {
      "schema": {
        "type": "object",
        "required": [
          "message",
          "code",
          "reason"
        ],
        "properties": {
          "code": {
            "description": "The HTTP status code",
            "type": "integer",
            "format": "int64",
            "x-go-name": "Code",
            "example": 400
          },
          "message": {
            "description": "Response content",
            "type": "string",
            "x-go-name": "Message",
            "example": "Invalid Data type"
          },
          "reason": {
            "description": "HTTP status text",
            "type": "string",
            "x-go-name": "Reason",
            "example": "Bad Request"
          }
        }
      }
    },
    "badRequestResponseWrapper": {
      "schema": {
        "type": "object",
        "required": [
          "message",
          "code",
          "reason"
        ],
        "properties": {
          "code": {
            "description": "The HTTP status code",
            "type": "integer",
            "format": "int64",
            "x-go-name": "Code",
            "example": 400
          },
          "message": {
            "description": "Response content",
            "type": "string",
            "x-go-name": "Message"
          },
          "reason": {
            "description": "HTTP status text",
            "type": "string",
            "x-go-name": "Reason",
            "example": "Bad Request"
          }
        }
      }
    },
    "okResponseWrapper": {
      "schema": {
        "type": "object",
        "required": [
          "message",
          "code",
          "reason"
        ],
        "properties": {
          "code": {
            "description": "The HTTP status code",
            "type": "integer",
            "format": "int64",
            "x-go-name": "Code",
            "example": 200
          },
          "message": {
            "description": "Response content",
            "type": "string",
            "x-go-name": "Message"
          },
          "reason": {
            "description": "HTTP status text",
            "type": "string",
            "x-go-name": "Reason",
            "example": "OK"
          }
        }
      }
    },
    "okloginResponseWrapper": {
      "schema": {
        "type": "object",
        "required": [
          "message",
          "code",
          "reason"
        ],
        "properties": {
          "code": {
            "description": "The HTTP status code",
            "type": "integer",
            "format": "int64",
            "x-go-name": "Code",
            "example": 200
          },
          "message": {
            "description": "Response content",
            "type": "string",
            "x-go-name": "Message",
            "example": "Verification code sent"
          },
          "reason": {
            "description": "HTTP status text",
            "type": "string",
            "x-go-name": "Reason",
            "example": "OK"
          }
        }
      }
    },
    "unprocessableEntityResponseWrapper": {
      "schema": {
        "type": "object",
        "required": [
          "message",
          "code",
          "reason"
        ],
        "properties": {
          "code": {
            "description": "The HTTP status code",
            "type": "integer",
            "format": "int64",
            "x-go-name": "Code",
            "example": 422
          },
          "message": {
            "description": "Response content",
            "type": "string",
            "x-go-name": "Message"
          },
          "reason": {
            "description": "HTTP status text",
            "type": "string",
            "x-go-name": "Reason",
            "example": "Unprocessable Entity"
          }
        }
      }
    },
    "unprocessableLoginResponseWrapper": {
      "schema": {
        "type": "object",
        "required": [
          "message",
          "code",
          "reason"
        ],
        "properties": {
          "code": {
            "description": "The HTTP status code",
            "type": "integer",
            "format": "int64",
            "x-go-name": "Code",
            "example": 422
          },
          "message": {
            "description": "Response content",
            "type": "string",
            "x-go-name": "Message",
            "example": "Invalid payload"
          },
          "reason": {
            "description": "HTTP status text",
            "type": "string",
            "x-go-name": "Reason",
            "example": "Unprocessable Entity"
          }
        }
      }
    }
  }
}